moved users to another directories
