+++ dook/api/admin/views.py
@@ -14,9 +14,9 @@ from dook.api.admin.serializers import (
     UserDetailSerializer,
 )
 from dook.api.news.filters import AdminNewsFilter
+from dook.api.users.filters import AdminUsersFilter
 from dook.core.news.models import News, SensitiveKeyword
-from dook.users.filters import AdminUsersFilter
-from dook.users.models import Invitation, User
+from dook.core.users.models import Invitation, User
 
 
 class ExpertListView(generics.ListAPIView):
