+++ dook/core/users/tokens.py
@@ -3,12 +3,12 @@ from django.contrib.auth.tokens import PasswordResetTokenGenerator
 from django.utils.encoding import force_bytes, force_str
 from django.utils.http import urlsafe_base64_decode, urlsafe_base64_encode
 
-from dook.users.exceptions import (
+from dook.core.users.exceptions import (
     EmailAlreadyConfirmedException,
     InvalidActivationUrlException,
     InvalidInviteTokenException,
 )
-from dook.users.models import User
+from dook.core.users.models import User
 
 
 def get_user_from_uid(uidb64):
