+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 from pytest_django.lazy_django import skip_if_no_django
 
-from dook.users.models import User
+from dook.core.users.models import User
 
 
 @pytest.fixture()
@@ -24,7 +24,6 @@ def api_client():
 def admin_api_client(db, admin_user):
     """A Django test client logged in as an admin user."""
     from rest_framework.test import APIClient
-    from rest_framework.authtoken.models import Token
 
     client = APIClient()
     client.force_authenticate(user=admin_user)
@@ -46,7 +45,6 @@ def default_user(db, django_user_model):
 @pytest.fixture()
 def authenticated_api_client(db, default_user):
     from rest_framework.test import APIClient
-    from rest_framework.authtoken.models import Token
 
     client = APIClient()
     client.force_authenticate(user=default_user)
