+++ tests/test_core/test_users/test_managers.py
@@ -6,9 +6,9 @@ from assertpy import assert_that
 from django.test import override_settings
 from django.utils import timezone
 
-from dook.users.constants import UserRoleType
-from dook.users.managers import UserQuerySet
-from dook.users.models import User, UserNews
+from dook.core.users.constants import UserRoleType
+from dook.core.users.managers import UserQuerySet
+from dook.core.users.models import User, UserNews
 from tests.factories.news import ExpertOpinionFactory, FactCheckerOpinionFactory
 from tests.factories.users import UserFactory, UserNewsFactory
